# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateConversationBody {
  title: string;
}

export interface SendMessageBody {
  content: string;
}

export interface ChangelogEntry {
  id: number;
  sortKey: string;
  date: string;
  category: string;
  summary: string;
  capabilities: string[];
  createdAt: string;
}

export interface ContentIdea {
  id: number;
  platform: string;
  theme: string;
  title: string;
  body: string;
  hashtags: string[];
  createdAt: string;
}

export interface GenerateContentIdeasBody {
  platform: string;
  theme: string;
  count?: number;
}

export type ListContentIdeasParams = {
  platform?: string;
  theme?: string;
};