# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListConversationsResponse = zod.array(
  ListConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get a conversation with its messages
 */
export const GetConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send a message and get a streamed response
 */
export const SendMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all knowledge changelog entries
 */
export const ListChangelogEntriesResponseItem = zod.object({
  id: zod.number(),
  sortKey: zod.string(),
  date: zod.string(),
  category: zod.string(),
  summary: zod.string(),
  capabilities: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListChangelogEntriesResponse = zod.array(
  ListChangelogEntriesResponseItem,
);

/**
 * @summary List all content ideas
 */
export const ListContentIdeasQueryParams = zod.object({
  platform: zod.coerce.string().optional(),
  theme: zod.coerce.string().optional(),
});

export const ListContentIdeasResponseItem = zod.object({
  id: zod.number(),
  platform: zod.string(),
  theme: zod.string(),
  title: zod.string(),
  body: zod.string(),
  hashtags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListContentIdeasResponse = zod.array(ListContentIdeasResponseItem);

/**
 * @summary Generate new AI content ideas
 */
export const GenerateContentIdeasBody = zod.object({
  platform: zod.string(),
  theme: zod.string(),
  count: zod.number().optional(),
});

/**
 * @summary Delete a content idea
 */
export const DeleteContentIdeaParams = zod.object({
  id: zod.coerce.number(),
});