# lib/api-zod/src/generated/types/changelogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangelogEntry {
  id: number;
  sortKey: string;
  date: string;
  category: string;
  summary: string;
  capabilities: string[];
  createdAt: Date;
}