# lib/api-zod/src/generated/types/contentIdea.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ContentIdea {
  id: number;
  platform: string;
  theme: string;
  title: string;
  body: string;
  hashtags: string[];
  createdAt: Date;
}