# lib/api-zod/src/generated/types/conversationWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  updatedAt: Date;
  messages: Message[];
}