# lib/api-zod/src/generated/types/generateContentIdeasBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtu Ferries Brand Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateContentIdeasBody {
  platform: string;
  theme: string;
  count?: number;
}